Create invert_chords.py
This is a small program that will invert guitar chords according to the fret you keep your capo in. Suppose the initial chord is C Major and you want to keep your capo in the second fret then you will enter the correct details and the program will suggest you the actual chord after the capo is placed.